fix(core): ScriptResult 整数优先 long/decimal 避免大整数精度丢失
脚本路径整数经 JsValue 时一律转 double 导致大整数（如数据库 ID 9007199254740993）
精度丢失。现已对齐 ParameterResolver.ResolveNumber 的策略：

- ToClr/ToJson/To<T> 数值分支：整数优先 long，超出 long 范围回退 decimal，
  仅非整数输出 double；ToJson 以整数字面量输出而非指数形式。
- 已预求值的 JSON 节点按 System.Text.Json 原类型直读，完整保留大整数精度
  （Jint 以 double 存储数字，2^53 以上整数无法经脚本执行路径还原，
  仅 JSON/预求值路径可保留）。

同步修正依赖脚本结果为 double 的现有断言，以及 Calculator 工具节点
对数值结果仅输出 double 的分支（现按实际数值类型输出）。

测试：新增大整数精度用例（脚本执行 + 预求值路径），保留 int/double 行为。

+++ tests/FlowEngine.Runtime.Tests/Plugins/CalculatorToolNodeTests.cs
@@ -39,7 +39,7 @@ public sealed class CalculatorToolNodeTests
 
         Assert.True(result.Success, result.Error?.Message);
         Assert.Equal("1 + 2", result.Output.Items[0].Data?["expression"]?.GetValue<string>());
-        Assert.Equal(3d, result.Output.Items[0].Data?["result"]?.GetValue<double>());
+        Assert.Equal(3, result.Output.Items[0].Data?["result"]?.GetValue<int>());
     }
 
     [Fact]
@@ -51,7 +51,7 @@ public sealed class CalculatorToolNodeTests
 
         Assert.True(result.Success, result.Error?.Message);
         Assert.Equal("10 / 2", result.Output.Items[0].Data?["expression"]?.GetValue<string>());
-        Assert.Equal(5d, result.Output.Items[0].Data?["result"]?.GetValue<double>());
+        Assert.Equal(5, result.Output.Items[0].Data?["result"]?.GetValue<int>());
     }
 
     [Fact]
@@ -63,7 +63,7 @@ public sealed class CalculatorToolNodeTests
 
         Assert.True(result.Success, result.Error?.Message);
         Assert.Equal("4 * 2", result.Output.Items[0].Data?["expression"]?.GetValue<string>());
-        Assert.Equal(8d, result.Output.Items[0].Data?["result"]?.GetValue<double>());
+        Assert.Equal(8, result.Output.Items[0].Data?["result"]?.GetValue<int>());
     }
 
     [Fact]
@@ -75,7 +75,7 @@ public sealed class CalculatorToolNodeTests
 
         Assert.True(result.Success, result.Error?.Message);
         Assert.Equal("7 - 3", result.Output.Items[0].Data?["expression"]?.GetValue<string>());
-        Assert.Equal(4d, result.Output.Items[0].Data?["result"]?.GetValue<double>());
+        Assert.Equal(4, result.Output.Items[0].Data?["result"]?.GetValue<int>());
     }
 
     [Fact]
@@ -88,7 +88,7 @@ public sealed class CalculatorToolNodeTests
 
         Assert.True(result.Success, result.Error?.Message);
         Assert.Equal("5", result.Output.Items[0].Data?["expression"]?.GetValue<string>());
-        Assert.Equal(5d, result.Output.Items[0].Data?["result"]?.GetValue<double>());
+        Assert.Equal(5, result.Output.Items[0].Data?["result"]?.GetValue<int>());
     }
 
     [Fact]
